Password history parsing
Password history from the notes section is parsed into the PwdSafe's password history field. To be considered a history item, a line in the notes should begin with "old: ", "was: ", or "password was: ". Current time is used as a password change timestamp unless a valid timestamp is specified in the notes after the password.